test-bot: handle deleted formulae.
When a formula is deleted the test bot run is currently a no-op. We may
want to add more here in future but to start with let's do a basic
`brew uses` check to ensure no formulae still depend on it.

Also, while we're here: pluralise `@modified_formula` for consistency.