test-bot: nuke etc/var before installation.
These directories are only ever needed at runtime and we're not running
anything except the main formula. The way bottle etc/var handling works
assumes a clean slate before `brew install` so this becomes considerably
more reliable when we just delete everything

Worst case this introduces some weird `brew test` failure in future that
mean we can `brew postinstall` (which, as-of, Homebrew/brew#2579 will
reinstall configuration files) on formulae as needed if we hit that.