Don't test dependents on audit/linkage/test failure.
We need to get on top of our failing CI problems. This will block us
doing dependent testing or shipping bottles for formulae where linkage
or tests fail on the changed formula. We will still allow this on
dependent formulae.

Also, run `brew audit` without Git, style (RuboCop) or online checks
in the tap syntax checking step. This is fast enough now (takes under
a minute on my machine) and these are the audit failures we should
never accept failures on.