Improve the readme
While this tool is used by core, the contributors who setup core and
maintain it are closely linked to homebrew project. As such they have in
depth knowledge of the domain and so are less likely to need this
document. There's also only one core (compared to the large number of
custom taps), as such this documentation may be helpful to core, but
it's not it's primary audience.

This document is written for the audience developers who don't have 100
stars (or whaver the required limit is) to get there formula into core,
and as such are creating a custom tap to distribute their application.

The key questions it aims to answer for those custom tap developers are:
* How do a test my new formula
* How to a automate the tests to that custom formula so people can
  contribute
* How do I build binaries for my formula so instals are fast

These are all tasks that testbot significantly makes simpler, but
pretty much no one I talk to knows about it because these great features
aren't easily discoverable.

There are also a number of idiosyncrasies that happen on GitHub Actions
that need to be documented for a smooth "oh wow it just works"
experience for first time users.

This is also our opportunity to state to the user what we expect you to
do, and, by ommission, what we don't expect, hopefully reducing
questions like #474, #302, and #251

This should also live in the main man pages for the `test-bot` tool.