lib/tests/formulae_dependents: skip unbottled source builds
We regularly see CI failures from attempted source builds of dependents
that aren't bottled (cf. any Go, Rust, or GHC PR). Let's avoid these by
not attempting these builds at all.

However, in the longer term, I believe it would be good to attempt these
source builds but reverse the failure condition: a build/test failure is
expected, so the test should pass when this happens, and fail when it
doesn't.

That would allow us to detect unbottled dependents that should be
bottled.

That change will take some work, however, so I've left a `TODO` in a
comment that I intend to revisit at some point.