lib/tests/formulae: tweak skipping formulae with unbottled deps
There is a corner case where an unbottled dependency is added to a
bottled formula. The current behaviour would silently skip the build,
which is undesirable. Let's retain the old behaviour for this case,
which is to attempt the build anyway (and most likely fail).

I'm not certain this is the correct behaviour yet, but the old one is,
in my opinion, still better than what is happening now.

See Homebrew/homebrew-core#88797.