lib/tests/formulae: relax test dep requirements when building bottles
Test-only dependencies have no impact on the bottle that is built, so
there is no need to check that these dependencies are bottled on the
current OS before a formula is bottled.

Let's relax the requirement that a test dependency is bottled for the
current OS to a check that a test dependency has a usable bottle. At
worst, the older bottle doesn't work and the test will fail. If it does
work, there's no reason for this dependency to block bottling.