Add a `--skip-online-checks` flag
Passing this flag to `test-bot` will omit the `--online` flag from
`brew audit`, and also skip `brew livecheck`.

This is useful for at least three cases:

1. Bottling on a new OS/architecture. These failures should probably not
   block this especially if the build/linkage/test and install audits
   succeeded.
2. Failures which seem to occur only in CI. (cf.
   Homebrew/homebrew-core#88329, Homebrew/homebrew-core#89042)
3. PRs which require revision bumps to multiple formulae. CI runs on
   these PRs often take a very long time, and this makes fixing these
   failures (which is sometimes not even possible) and then re-running
   CI very impractical.

The last point is a particularly sore pain point for me, because I find
few things more frustrating than waiting a long time for CI to run only
to find that dependent tests have been blocked because livecheck failed,
or because the audit failed because the stable tarball was not reachable
(despite the source build completing successfully).