lib/tests/formulae: demote more unbottled failures to warnings
Let's demote another class of CI failures from unbottled formulae to
warnings: attempting to build a formula with a dependency that is also
in the `@testing_formulae` array. We see this in Homebrew/homebrew-core#89193.

It would be nice to be able to restrict the failures we ignore to only
the unbottled ones, but we need to pass the `changed_dependencies` in
its entirety to `brew install` in order for `brew` to work out the
correct order to install these dependencies.