lib/step: enable annotations for all errors
This implements a way to produce annotations for all errors in a
`test-bot` run. It requires baking in some assumptions that we are
testing formulae into a `Step`. In exchange, we get annotations on
relevant formula files.

Additionally, a `brew install` or `brew test` failure will place that
annotation on the `install` method or `test` block respectively.

To enable annotations on formula files, we add a `named_args` keyword
argument to the `test` method. Formula names (or formulae) should be
passed as the value of `named_args` for annotations to be emitted.

I've also moved the `ignore` logic into the `Step` class, so there
should be no need to call `Step#ignore` manually after a `test` if we
wish to ignore errors from that step. Instead, we pass
`ignore_failures: true`.