formulae_dependents: build more dependents from source
If the `--build-dependents-from-source` flag is passed, then all
dependents are built from source. This is the current behaviour.

We change behaviour in two ways:
1. Build unbottled dependents from source if all its dependencies are
   either bottled, or were built earlier in the `test-bot` run.
2. Build Linux-only formulae from source too (either with the
   appropriate flag or when the above condition is true) instead of
   always skipping source builds on Linux.

Building unbottled dependents will be useful for detecting dependents
which should be bottled after #714.

This change should also allow us to clean up a bit of code in the
`#dependent_formulae!` method. I've omitted that for now since I have a
pending PR (#717) which introduces conflicting changes.