Combine duplicated logic
On macOS, all formulae are possible `source_dependents`. On Linux, only
Linux-only formulae are possible `source_dependents`.

From the formulae satisfying the above criteria, we insert only the ones
with useable dependencies into the `source_dependents` array. If the
`--build-dependents-from-source` flag is passed, then all such
dependents are placed into the `source_dependents` array. Otherwise, we
select only the ones that are not bottled on the current OS.