formulae: check deps against `testing_formulae` instead
This is a better fix for the issue I tried to fix in #733.

Checking dependents against `@new_formula` allows you to declare a
formula dependency on a new formula added in the same PR. The same thing
is achieved by checking `@testing_formulae` instead, but also accounts
for the case where a PR fixes a previously-unbottled formula `x`, and
also makes `x` a dependency of another formula.