formulae_dependents: build dependents from source only on request
I trialled this behaviour in Homebrew/homebrew-core#89857, and it adds a
significant amount of time to the CI run. I'm still interested in
automating the detection of formulae that should be bottled, but I think
it requires a more nuanced approach.

In the meantime, let's restore the old behaviour that dependents are
built from source only when requested via a flag.

Closes #731.