Fix bottling dependents of new formulae
This is still happening despite my previous attempts to fix it. The
issue is that a formula is removed from the `@testing_formulae` array
once its bottle is successfully built, so I can't check this array to
determine whether a formula was successfully built earlier in the run.

Let's fix that by keeping track of formulae with built bottles
separately.

I tested this a lot more carefully than my previous attempts, so
hopefully this should do it now.

Needed for Homebrew/homebrew-core#92329, Homebrew/homebrew-core#93167,
and Homebrew/homebrew-core#92024.