Add `--skip-recursive-build-dependents` flag
This is based on the discussions at Homebrew/brew#12455.

My intention is for workflows in Homebrew/core to always pass
`--skip-recursive-build-dependents`, perhaps with a label that causes
the workflow to omit the flag.

I'm tempted to not make this an option at all, but this changes testing
behaviour in external taps too, and I'm a little wary about forcing
those changes when they haven't been asked for.