Handle livecheck error conditions
The existing `livecheck_info.first["version"]["newer_than_upstream"]`
line assumes that the first object in the JSON array will contain a
`"version"` hash. This is true when the check is successful but a
`"version"` hash isn't present when an error is encountered:

```
{
  "formula": "example",
  "status": "error",
  "messages": [
    "curl: (6) Could not resolve host: example.com"
  ],
  "meta": {
    "livecheckable": true
  }
}
```

`test-bot` doesn't account for a `nil` `"version"` value and this is
currently producing an "Error: undefined method `[]' for
nil:NilClass" error on CI. This commit addresses the issue by
accounting for a `nil` `"version"` value.