formulae_dependents: prune more recursive dependencies
The dependencies of build dependencies do not inherit the `:build` tag
when expanding via `#recursive_dependencies`. This results in `test-bot`
testing build-only dependents when it doesn't need to. See
Homebrew/homebrew-core#109250.

Let's fix that by pruning the dep tree whenever we encounter a build
dependency while expanding a dependent's dep tree. In the case of
`ninja`, this will result in `test-bot` correctly identifying that it
only needs to test formulae with a runtime dependency on `ninja`.