tests/formulae: downgrade optional steps to warnings
Some of the `brew fetch` and `brew install --only-dependencies` that we
do will be repeated if necessary when we `brew install` the formula
being tested.

Thus, it makes sense for failures of these steps to be warnings rather
than errors. This will help minimise the noise in CI runs where a `brew fetch`
failed but the actual build and test succeeded.