test_bot: set `HOMEBREW_EVAL_ALL`.
`test-bot` needs to be able to evaluate all formulae to:
1. run `tap-syntax`
2. determine dependents

There may be other places it needs this, but those are just the ones off
the top of my head.

To do this, we need to set `HOMEBREW_EVAL_ALL` after #13790.

Fixes:

    Error: Calling brew uses --all is deprecated! Use brew uses --eval-all or HOMEBREW_EVAL_ALL instead.

See https://github.com/Homebrew/homebrew-core/runs/8190122846?check_suite_focus=true#step:10:22.