Add --skip-livecheck option
Sometimes livecheck will work for a formula locally but fail on CI
(intermittently or consistently) and this can force us to merge a PR
with failing CI when there isn't an obvious solution. Ideally, we
should identify the underlying source of the failure and resolve it
but a `--skip-livecheck` option may be useful when that's not
immediately possible or the failure point is out of our control
(e.g., an upstream server issue).