Fix passing of `skipped_or_failed_formulae` to `test_deps`
The list of skipped or failed formulae is currently not being passed
correctly to the separate `test_deps` job after this was split in
Homebrew/homebrew-core#125595.

The current way of passing this list through step outputs is tricky to
get right with matrix jobs. However, we already pass a `bottles`
artifact between these jobs, so we can use that to record the
information that we need to pass on to the `test_deps` job.

When this is merged, we only need to remove the
`--skipped-or-failed-formulae` argument in the `brew test-bot`
invocation in the `test_deps` job.