formulae_dependents: dump list of skipped formulae
This will be useful for debugging purposes, in the same way that we dump
the list of dependents for each testing formula.

Also, fall through to an empty array if we can't determine the skipped
or failed formulae through normal means. This allows dependent testing
to continue assuming that none of the testing formulae failed.