Improve `#cleanup_during!`
On some long-running CI runs, you can see multiple consecutive calls to
`#cleanup_during!` in the CI log. This indicates that the runner is
close to running out of space, but deleting `HOMEBREW_CACHE` isn't quite
going far enough in clearing it out.

Let's try to improve that by uninstalling installed formulae that are
safe to uninstall on the second call to `#cleanup_during!` onwards.