cmd/test-bot: always set `--local` in GitHub Actions
Passing `--local` depending on whether the runner is self-hosted or not
complicates our handling of log directories in Homebrew/core.

Always setting it ensures that we can always look in the same place
relative to the bottles directory for the logs, and will allow us to
simplify some code in our workflows.

This also fix log uploads when using GitHub macOS runners for testing.