test_formulae: improve result caching
Our current method of using artifacts as a cache breaks when a PR is
pushed to multiple times in quick succession.

This is because we use the SHA1 prior to the most recent push to look up
artifacts, but there will be no useable artifacts to find there after
pushing to a PR that has recently been pushed to.

We can improve this by looking for an even earlier SHA1 to try to find
useable artifacts from.

This is a companion to Homebrew/homebrew-core#132093.