test_formulae: avoid repeated downloads of the same artifact
We can correctly follow the trail left by the PR `event.json`s but still
end up in an infinite loop for some reason.

Let's avoid this by giving up as soon as we find we're trying to
download an artifact we've already seen before.

Fixes a CI failure at Homebrew/homebrew-core#134939.