Add option to skip stable_version audit
I recently split the stable version formula audit into a separate
method that can be skipped using `--except`. This adds a
`--skip-stable-version-audit` option to test-bot that passes
`--except=stable_version` to `brew audit`. This will be controlled in
homebrew/core using a `CI-version-downgrade` label (for when we need
a downgrade PR to run on CI).