formulae: specify that linter should only run on named formula
This reduces the scope of this audit to the expected use case.
We are only trying to lint a single formula file here. This can
become ambiguous since `brew style` doesn't know automatically
if you are trying to lint a path, tap, cask or formula. We saw
problems tangentially related to this in `homebrew/core` recently
that were also based on changes to the cask loader which seem
to have changed the default cask path in certain situations but
this change allows us to work around those problems entirely.