test_formulae: preserve logs in `cleanup_during!`
On Linux by default, `HOMEBREW_LOGS` (`~/.cache/Homebrew/Logs`) is a
subdirectory of `HOMEBREW_CACHE` (`~/.cache/Homebrew`). This means that
`cleanup_during!` deletes the logs for the already-tested formulae,
which is not ideal because these logs are to be uploaded as an artifact.
These logs don't usually take up as much space as the cached downloads,
and can be valuable for debugging, especially when a PR involves too
many formulae that GitHub Actions cannot show all the logs.

Instead, we can move the logs to a temporary directory, and restore them
after the cleanup.