Update usage of `Tap#formula_file?`
Currently this method takes either a String or a Pathname which
is fine since we were passing in a Pathname. I'm planning on
changing that behavior in new PR so that it will only accept
the String class going forward so I'm updating this to use the
String class here. This should be completely backwards compatible.