Add `--tested-formulae` flag
This is needed for Homebrew/homebrew-core#182696.

The problem that lead to the revert in Homebrew/homebrew-core#182754 is
due to `test-bot` not realising that there are dependents we had already
tested in the earlier formulae build job that do not need to be tested
anymore.

We can fix this by making sure `test-bot` knows all the formulae we had
previously tested passed via a `--tested-formulae` flag.

For clarity, let's rename the old `@tested_formulae` ivar to
`@testing_formulae_with_tested_dependents`.