Include first error message in GitHub annotation
The first error message is often found before the last 256 lines, so
isn't included in the annotation.

Let's adjust `#truncate_output` so it includes this error message if
found instead. Searching for error messages is a little naive, but it
matches the way `clang`, `gcc`, and `rustc` print error messages. We can
make this more sophisticated later if desired.