Truncate annotation bodies to 4KB
GitHub Actions seems to truncate our annotations to 4KB, so let's adjust
to truncating at 4KB instead.

Since 4KB isn't a lot, let's drop the line count handling, and reduce
the number of context lines we have for error messages.

While we're here, let's also avoid truncating the start of our output in
the middle of a line, since that makes the first line of our annotation
less intelligible.