formulae_dependents: improve `texlive` handling
The `texlive` bottle is gigantic and has about a thousand dependencies.
This tends to make GitHub-hosted runners run out of storage by the time
they get around to testing it, particularly since it comes late in the
alphabet (and we test dependents in alphabetical order).

To try to minimise usage of our self-hosted runner, let's test `texlive`
as early as possible to increase the chances of the runner having
sufficient storage to be able to test `texlive`.