test_formulae: relax bottle requirement for test deps of `:all` bottles
`test-bot` is erroneously skipping builds for dependents of formulae
with `:all` bottles when the formula with an `:all` bottle has a test
dependency that isn't bottled.

See, for example:

https://github.com/Homebrew/homebrew-core/actions/runs/11278840144/job/31382054266#step:3:33

In the above example, `sfcgal` depends on `cgal`, and `cgal` has an
`:all` bottle and a `:test` dependency on `qt`, which is not yet bottled
on Sequoia. We should be ok with building a bottle for `sfcgal` in this
case.