tests/formulae_dependents: another `--ignore-dependencies` fix
One more (and hopefully last) fix like #1305 and #1306. When we test
dependents of multiple formulae, a dependent of one formula being tested
can happen to be a dependency of another formula being tested. In this
case, we need the `--ignore-dependencies` option to force removal of the
dependent/dependency formula. When testing the dependents of the removed
formula, `brew install --only-dependencies` ensures that it is
reinstalled.

Fixes failure seen at https://github.com/Homebrew/homebrew-core/actions/runs/11923246241/job/33242018653?pr=198240.