formulae: use livecheck --autobump option
We recently modified the behavior of `brew livecheck` to skip
packages in the tap's autobump list, so this change is causing
autobumped formulae to fail the `brew livecheck` part of CI testing.
This adds the `--autobump` option to the `brew livecheck` call, which
will ensure that livecheck doesn't skip autobumped formulae.