don't link dependencies with conflicts
The new anyzig formula depends on the zig formula, but, it also
conflicts_with it because they both install `zig` binaries. This works
fine with brew but test-bot is failing because it's linking the dependency
before installing, even with the conflict. The fix is to disable linking
dependencies that are also marked as conflicts.